@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -29,10 +29,17 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   test: ^1.16.5
+  moor: # use the latest version
+  sqlite3_flutter_libs: # Also use the latest version.
+  path_provider:
+  path:
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+
+  moor_generator: # use the latest version
+  build_runner:
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
