@@ -1,7 +1,6 @@
-import 'package:flutter/foundation.dart';
 
 abstract class Indexable {
   final Symbol id;
 
-  const Indexable({@required this.id});
+  const Indexable({required this.id});
 }
