@@ -7,7 +7,7 @@ import 'unit.dart';
 class MeasurementType implements Indexable {
   final Symbol id;
   final Units units;
-  const MeasurementType({this.id, this.units});
+  const MeasurementType({required this.id, required this.units});
 
   static String format(Map<MeasurementType, Quantity> measurements) {
     return "measurements ${measurements.entries.map((e) => "${TL8(e.key.id)} (${TL8(e.key.units.id)}): ${e.value.amount} ${TL8(e.value.units.id)}").join("; ")}";
